gen-ai/Assistants/travel_planner/requirements.txt
openai==1.30.1
python-dotenv